# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple API for managing player profiles and achievements",
        "title": "Player Profile API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/players": {
            "get": {
                "description": "Get all players with pagination, by default page is 1 and pageSize is 10",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Player"
                ],
                "summary": "Get all players",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new player profile with the input payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Player"
                ],
                "summary": "Create a new player profile",
                "parameters": [
                    {
                        "description": "Create Player Profile Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreatePlayerProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/players/{playerID}": {
            "get": {
                "description": "Get player by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Player"
                ],
                "summary": "Get player by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Player ID",
                        "name": "playerID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update player by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Player"
                ],
                "summary": "Update player by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Player ID",
                        "name": "playerID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Player Profile Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdatePlayerProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete player by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Player"
                ],
                "summary": "Delete player by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Player ID",
                        "name": "playerID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/players/{playerID}/achievements": {
            "get": {
                "description": "Get player with achievements by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Player"
                ],
                "summary": "Get player with achievements by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Player ID",
                        "name": "playerID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.CreatePlayerProfileRequest": {
            "type": "object",
            "required": [
                "avatar",
                "experience",
                "level",
                "nickname",
                "point",
                "user_id"
            ],
            "properties": {
                "avatar": {
                    "description": "Player avatar",
                    "type": "string"
                },
                "experience": {
                    "description": "Player experience",
                    "type": "integer"
                },
                "level": {
                    "description": "Player level",
                    "type": "integer"
                },
                "nickname": {
                    "description": "Player nickname",
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 3
                },
                "point": {
                    "description": "Player points",
                    "type": "integer"
                },
                "user_id": {
                    "description": "User ID (foreign key) in the database",
                    "type": "integer"
                }
            }
        },
...
        "request.UpdatePlayerProfileRequest": {
            "type": "object",
            "required": [
                "avatar",
                "experience",
                "level",
                "nickname",
                "points"
            ],
            "properties": {
                "avatar": {
                    "description": "Player avatar",
                    "type": "string"
                },
                "experience": {
                    "description": "Player experience",
                    "type": "integer"
                },
                "level": {
                    "description": "Player level",
                    "type": "integer"
                },
                "nickname": {
                    "description": "Player nickname",
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 3
                },
                "points": {
                    "description": "Player points",
                    "type": "integer"
                }
            }
        },
...
        "response.BaseResponse": {
            "description": "Base response structure",
            "type": "object",
            "properties": {
                "code": {
                    "description": "HTTP status code of the response",
                    "type": "integer",
                    "x-order": "0",
                    "example": 200
                },
                "status": {
                    "description": "Status of the response",
                    "type": "string",
                    "x-order": "1",
                    "example": "success"
                },
                "message": {
                    "description": "Message of the response",
                    "type": "string",
                    "x-order": "2",
                    "example": "Operation completed successfully"
                },
                "data": {
                    "description": "Data payload of the response",
                    "x-order": "3"
                }
            }
        }
    }
}